# api/openapi.yaml
openapi: 3.1.0
info:
  title: Omahab Control API
  description: |
    Authenticated control-plane API for omahabd. All routes are under /api/v1 except
    /up. Bearer authentication is required except for /up and the email webhook which
    uses a separate HMAC authentication path. Mutating requests require Content-Type
    application/json. Unknown JSON fields are rejected. Request bodies are limited
    (1 MiB JSON, 5 MiB email webhook) and per-request timeouts are enforced. SSE
    streaming at /api/v1/events/stream supports Last-Event-ID replay and heartbeat
    comments. Secret and provider-credential reads return metadata only.
  version: 0.1.0
  contact:
    name: Omahab
servers:
  - url: http://localhost:8484
    description: omahabd default listen address
security:
  - bearerAuth: []
paths:
...
  /api/v1/events/stream:
    get:
      summary: Stream events via SSE
      description: |
        Server-Sent Events stream. Honors Last-Event-ID header (or ?lastEventId query) for replay.
        Sends heartbeat comments every 15s. Respects client cancellation and server shutdown.
      operationId: streamEvents
      parameters:
        - name: lastEventId
          in: query
          schema: { type: string }
          required: false
        - name: Last-Event-ID
          in: header
          schema: { type: string }
          required: false
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Bearer token verified via constant-time SHA256 comparison. Required for all /api/v1 routes except /up and /api/v1/email/ingest.
    hmacAuth:
      type: apiKey
      in: header
      name: X-Omahab-Signature
      description: Worker v1 canonical HMAC-SHA256, optionally prefixed with sha256=. Metadata mirror headers are also required.
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
      headers:
        X-Request-ID: { schema: { type: string } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: not_found }
            message: { type: string }
...